+++ R/load_forecasts_zoltar.R
@@ -1,6 +1,6 @@
 #' Load all available forecasts submitted on `forecast_dates` from Zoltar.
 #'
-#' This function will throw a warning and return an empty data frame when
+#' This function will throw a warning and return empty data.frame when
 #' no forecasts are submitted on any dates in `forecast_dates` for selected `models`,
 #' `locations`, `types` and `target`.
 #'
@@ -67,14 +67,18 @@ load_forecasts_zoltar <- function(models = NULL,
   `%dopar%` <- foreach::`%dopar%`
 
   if (!is.null(forecast_dates)) {
-    
-    if (is.null(models)){
+    if (is.null(models)) {
       models <- all_models$model_abbr
       models <- sort(models, method = "radix")
     }
-    
+
     # set 2 workers
-    cl <- parallel::makeCluster(2, setup_strategy = "sequential")
+    if (verbose) {
+      cl <- parallel::makeCluster(2, setup_strategy = "sequential", outfile = "")
+    } else {
+      cl <- parallel::makeCluster(2, setup_strategy = "sequential")
+    }
+
     doParallel::registerDoParallel(cl)
     forecasts <- foreach::foreach(i = 1:length(models), .combine = rbind) %dopar% {
       curr_model <- models[i]
@@ -90,17 +94,20 @@ load_forecasts_zoltar <- function(models = NULL,
       latest_dates <- purrr::map(
         forecast_dates,
         function(a_list) {
-          max(intersect(
+          intersection <- intersect(
             as.character(a_list),
             as.character(model_forecasts_history)
-          ))
+          )
+          if (length(intersection) > 0) {
+            max(intersection)
+          }
         }
       )
 
       # unlist and drop duplicates
       latest_dates <- unique(unlist(latest_dates, use.names = FALSE))
       latest_dates <- latest_dates[!is.na(latest_dates)]
-      
+
       if (length(latest_dates) != 0) {
         forecast <- zoltr::do_zoltar_query(
           zoltar_connection = zoltar_connection,
@@ -114,11 +121,25 @@ load_forecasts_zoltar <- function(models = NULL,
           verbose = verbose,
           as_of = date_to_datetime(as_of, hub)
         )
-  
-        forecast <- reformat_forecasts(forecast)
+
+        if (nrow(forecast) == 0 & verbose) {
+          warning(paste0(
+            "Warning in load_forecasts_zoltar: Forecasts are not available for current model ",
+            curr_model
+          ))
+        }
+        forecast <- reformat_forecasts(forecast, verbose)
+      } else {
+        if (verbose) {
+          warning(paste0(
+            "Warning in load_forecasts_zoltar: No available forecast dates for current model ",
+            curr_model
+          ))
+        }
       }
     }
     # shut down workers
+    doParallel::stopImplicitCluster()
     parallel::stopCluster(cl)
   } else {
     # load forecasts submitted on all dates
@@ -134,21 +155,22 @@ load_forecasts_zoltar <- function(models = NULL,
       verbose = verbose,
       as_of = date_to_datetime(as_of, hub)
     )
-    forecasts <- reformat_forecasts(forecasts)
+    forecasts <- reformat_forecasts(forecasts, verbose)
   }
-  
-  if (!is.null(forecasts)){
-    if (nrow(forecasts) > 0) {
+
+  if (!is.null(forecasts) && !is.null(nrow(forecasts))) {
+    if (nrow(forecasts) > 0 && "location" %in% colnames(forecasts)) {
       # append location, population information
       forecasts <- forecasts %>%
         join_with_hub_locations(hub = hub)
     } else {
+      forecasts <- data.frame()
       warning("Warning in load_forecasts_zoltar: Forecasts are not available.\n Please check your parameters.")
     }
-  }else {
+  } else {
+    forecasts <- data.frame()
     warning("Warning in load_forecasts_zoltar: Forecasts are not available.\n Please check your parameters.")
   }
-  
   return(forecasts)
 }
 
@@ -159,14 +181,17 @@ load_forecasts_zoltar <- function(models = NULL,
 #' no forecasts are available in `zoltar_query_result`.
 #'
 #' @param zoltar_query_result dataframe returned by [zoltr::do_zoltar_query]
+#' @param verbose logical for printing error messages when checking `zoltar_query_result`. Default to `TRUE`.
 #'
 #' @return data.frame with columns `model`, `forecast_date`, `location`, `horizon`,
 #' `temporal_resolution`, `target_variable`, `target_end_date`, `type`, `quantile`, `value`
 #'
 #' @export
-reformat_forecasts <- function(zoltar_query_result) {
+reformat_forecasts <- function(zoltar_query_result, verbose = TRUE) {
   if (nrow(zoltar_query_result) == 0) {
-    warning("Warning in reformat_forecasts: Forecasts are not available.\n Please check your parameters.")
+    if (verbose) {
+      warning("Warning in reformat_forecasts: Forecasts are not available.\n Please check your parameters.")
+    }
     # convert value column to double and select columns
     zoltar_query_result <- zoltar_query_result %>%
       dplyr::mutate(value = as.double(value))